requirements.txt
python-dotenv
pandas
matplotlib
seaborn
tweepy
nltk
sastrawi
psycopg2-binary
